+++ tests/test_pattern.py
@@ -12,6 +12,7 @@ from numpy.random import PCG64, Generator
 from graphix.branch_selector import ConstBranchSelector, FixedBranchSelector
 from graphix.clifford import Clifford
 from graphix.command import C, Command, CommandKind, E, M, N, X, Z
+from graphix.flow.core import XZCorrections
 from graphix.flow.exceptions import (
     FlowError,
 )
@@ -1038,6 +1039,48 @@ class TestPattern:
         s_test = p_test.simulate_pattern(input_state=PlanarState(Plane.XZ, alpha))
 
         assert np.abs(np.dot(s_ref.flatten().conjugate(), s_test.flatten())) == pytest.approx(1)
+
+    # Extract xz-corrections from random circuits
+    @pytest.mark.parametrize("jumps", range(1, 11))
+    def test_extract_xzc_rnd_circuit(self, fx_bg: PCG64, jumps: int) -> None:
+        rng = Generator(fx_bg.jumped(jumps))
+        nqubits = 2
+        depth = 2
+        circuit_1 = rand_circuit(nqubits, depth, rng, use_ccx=False)
+        p_ref = circuit_1.transpile().pattern
+        xzc = p_ref.extract_xzcorrections()
+        xzc.check_well_formed()
+        p_test = xzc.to_pattern()
+
+        for p in [p_ref, p_test]:
+            p.remove_input_nodes()
+            p.perform_pauli_measurements()
+
+        s_ref = p_ref.simulate_pattern(rng=rng)
+        s_test = p_test.simulate_pattern(rng=rng)
+        assert np.abs(np.dot(s_ref.flatten().conjugate(), s_test.flatten())) == pytest.approx(1)
+
+    def test_extract_xzc_empty_domains(self) -> None:
+        p = Pattern(input_nodes=[0], cmds=[N(1), E((0, 1))])
+        xzc = p.extract_xzcorrections()
+        assert xzc.x_corrections == {}
+        assert xzc.z_corrections == {}
+        assert xzc.partial_order_layers == (frozenset({0, 1}),)
+
+    def test_extract_xzc_easy_example(self) -> None:
+        pattern = Pattern(
+            input_nodes=list(range(5)),
+            cmds=[M(0), M(1), M(2, s_domain={0}, t_domain={1}), X(3, domain={2}), M(3), Z(4, domain={3})],
+        )
+
+        xzc = pattern.extract_xzcorrections()
+        xzc_ref = XZCorrections.from_measured_nodes_mapping(
+            pattern.extract_opengraph(), x_corrections={0: {2}, 2: {3}}, z_corrections={1: {2}, 3: {4}}
+        )
+        assert xzc.og.isclose(xzc_ref.og)
+        assert xzc.x_corrections == xzc_ref.x_corrections
+        assert xzc.z_corrections == xzc_ref.z_corrections
+        assert xzc.partial_order_layers == xzc_ref.partial_order_layers
 
 
 def cp(circuit: Circuit, theta: Angle, control: int, target: int) -> None: